utilities/requirements.txt
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
google-genai==1.60.0
google-auth==2.47.0
httpx==0.28.1
langchain-openai==1.1.7
langchain-pinecone==0.2.13
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
python-dotenv==1.2.1
pinecone==7.3.0
flask==3.1.2
flask-cors==6.0.2
flask-limiter==4.1.1
python-dotenv